Move recommended videos to Research page and update Home page
